requirements.txt
# This file was autogenerated by uv via the following command:
#    uv pip compile requirements.in --output-file requirements.txt
anyio==4.8.0
    # via starlette
argparse==1.4.0
    # via -r requirements.in
cachetools==5.5.1
    # via google-auth
certifi==2025.1.31
    # via requests
charset-normalizer==3.4.1
    # via requests
click==8.1.7
    # via uvicorn
coverage==7.6.1
    # via pytest-cov
dill==0.3.9
    # via -r requirements.in
fastapi==0.101.0
    # via -r requirements.in
google-api-core==2.24.1
    # via google-cloud-pubsub
google-auth==2.38.0
    # via
    #   google-api-core
    #   google-cloud-pubsub
google-cloud-pubsub==2.21.1
    # via -r requirements.in
googleapis-common-protos==1.67.0
    # via
    #   google-api-core
    #   grpc-google-iam-v1
    #   grpcio-status
grpc-google-iam-v1==0.14.0
    # via google-cloud-pubsub
grpcio==1.70.0
    # via
    #   google-api-core
    #   google-cloud-pubsub
    #   googleapis-common-protos
    #   grpc-google-iam-v1
    #   grpcio-status
grpcio-status==1.62.3
    # via
    #   google-api-core
    #   google-cloud-pubsub
h11==0.14.0
    # via uvicorn
idna==3.10
    # via
    #   anyio
    #   requests
iniconfig==2.0.0
    # via pytest
packaging==24.1
    # via pytest
pluggy==1.5.0
    # via pytest
proto-plus==1.26.0
    # via
    #   google-api-core
    #   google-cloud-pubsub
protobuf==4.25.6
    # via
    #   google-api-core
    #   google-cloud-pubsub
    #   googleapis-common-protos
    #   grpc-google-iam-v1
    #   grpcio-status
    #   proto-plus
psycopg2-binary==2.9.7
    # via -r requirements.in
pyasn1==0.6.1
    # via
    #   pyasn1-modules
    #   rsa
pyasn1-modules==0.4.1
    # via google-auth
pydantic==1.10.13
    # via fastapi
pytest==7.4.0
    # via
    #   -r requirements.in
    #   pytest-cov
pytest-cov==4.1.0
    # via -r requirements.in
python-dotenv==1.0.1
    # via -r requirements.in
python-multipart==0.0.6
    # via -r requirements.in
requests==2.32.3
    # via google-api-core
rsa==4.9
    # via google-auth
sniffio==1.3.1
    # via anyio
sqlalchemy==2.0.19
    # via -r requirements.in
starlette==0.27.0
    # via fastapi
typing-extensions==4.12.2
    # via
    #   anyio
    #   fastapi
    #   pydantic
    #   sqlalchemy
urllib3==2.3.0
    # via requests
uvicorn==0.23.2
    # via -r requirements.in
